cvr_xbrl_app/requirements.txt
streamlit
requests
beautifulsoup4
lxml
arelle-release